+++ src/index.js
@@ -4,4 +4,5 @@ module.exports.utils = {
   report: require("./utils/report"),
   ruleMessages: require("./utils/ruleMessages"),
   styleSearch: require("./utils/styleSearch"),
+  ruleTester: require("./testUtils/ruleTester"),
 }