+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Head
+* Added: `ruleTester` to `stylelint.utils` for use by authors of custom rules.
+
 # 0.3.2
 * Fixed: `hierarchicalSelectors` bug in `indentation` rule.