+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added: `rule-no-shorthand-property-overrides` rule.
 * Added: `ruleTester` to `stylelint.utils` for use by authors of custom rules.
+* Removed: `jsesc` devDependency.
 
 # 0.3.2